Started moving vue login to rest api
#story[167] #pair[aho113, jsv22]

+++ server/src/main/java/com/springvuegradle/auth/AuthInterceptor.java
@@ -10,6 +10,7 @@ import javax.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
+import org.springframework.web.servlet.ModelAndView;
 
 import com.springvuegradle.model.data.Session;
 import com.springvuegradle.model.data.User;
@@ -30,6 +31,20 @@ public class AuthInterceptor implements HandlerInterceptor {
 	@Override
 	public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler)
 			throws Exception {
+
+		if (!response.containsHeader("Access-Control-Allow-Origin")) {
+			response.addHeader("Access-Control-Allow-Origin", "http://localhost:9500");
+		}
+		if (!response.containsHeader("Access-Control-Allow-Methods")) {
+			response.addHeader("Access-Control-Allow-Methods", "GET, PUT, POST, DELETE, OPTIONS");
+		}
+		if (!response.containsHeader("Access-Control-Allow-Headers")) {
+			response.addHeader("Access-Control-Allow-Headers", "content-type,DNT");
+		}
+		if (!response.containsHeader("Access-Control-Allow-Credentials")) {
+			response.addHeader("Access-Control-Allow-Credentials", "true");
+		}
+
 		boolean needsAuthentication = true;
 		for (String endpoint : this.nonAuthenticatedEndpoints) {
 			if (request.getRequestURI().equalsIgnoreCase(endpoint)) {
@@ -40,34 +55,39 @@ public class AuthInterceptor implements HandlerInterceptor {
 
 		if (needsAuthentication) {
 			boolean authenticated = false;
-			for (Cookie cookie : request.getCookies()) {
-				if (cookie.getName().equalsIgnoreCase("sessionid")) {
-					String sessionId = cookie.getValue();
-					if (sessionRepo.existsById(sessionId)) {
-						Session session = sessionRepo.findById(sessionId).get();
-						if (session.getExpiry().isBefore(Instant.now().atOffset(ZoneOffset.UTC))) {
-							sessionRepo.delete(session);
-						} else {
-							authenticated = true;
-							User user = session.getUser();
-							long profileId = user.getUserId();
-							if (adminRepo.existsById(user.getUserId())) {
-								profileId = -1;
+			if (request.getCookies() != null)
+				for (Cookie cookie : request.getCookies()) {
+					if (cookie.getName().equalsIgnoreCase("sessionid")) {
+						String sessionId = cookie.getValue();
+						if (sessionRepo.existsById(sessionId)) {
+							Session session = sessionRepo.findById(sessionId).get();
+							if (session.getExpiry().isBefore(Instant.now().atOffset(ZoneOffset.UTC))) {
+								sessionRepo.delete(session);
+							} else {
+								authenticated = true;
+								User user = session.getUser();
+								long profileId = user.getUserId();
+								if (adminRepo.existsById(user.getUserId())) {
+									profileId = -1;
+								}
+								request.setAttribute("authenticateduser", user);
+								request.setAttribute("authenticatedid", profileId);
 							}
-							request.setAttribute("authenticateduser", user);
-							request.setAttribute("authenticatedid", profileId);
 						}
 					}
 				}
-			}
 
 			if (authenticated) {
 				return true;
 			} else {
-				response.setStatus(401);
-				response.setContentType("application/json");
-				response.getOutputStream().write("{\"error\": \"Auth you are not\"}".getBytes());
-				return false;
+				if (!request.getMethod().equals("OPTIONS")) {
+					response.setStatus(401);
+					response.setContentType("application/json");
+					response.getOutputStream().write("{\"error\": \"Auth you are not\"}".getBytes());
+					return false;
+				} else {
+					return true;
+				}
 			}
 		} else {
 			return true;
